[ADD] Comment control middlewares

+++ src/api/v1/Comment/Comment.ctrl/deleteComment.ts
@@ -1,18 +1,21 @@
 import { Request, Response } from 'express';
+import ColorConsole from '../../../../lib/ColorConsole';
+import { handleFailed, handleSuccess } from '../../../../lib/Response';
 import { Repository, getRepository } from 'typeorm';
 import { Comment } from '../../../../entity/Comment';
+import { User } from 'entity/User';
 
 export default async (request: Request, response: Response) => {
 	try {
 		const { idx } = request.query;
+		const user: User = request.user;
 
 		const commentRepository: Repository<Comment> = getRepository(Comment);
 
-		if (isNaN(idx)) {
-			return response.status(400).json({
-				status: 400,
-				message: '검증 오류입니다.',
-			});
+		if (!Number.isInteger(idx)) {
+			ColorConsole.red(`[ERROR 400] 검증 오류입니다.`);
+			handleFailed(response, 400, '검증 오류입니다.');
+			return;
 		}
 
 		const findComment: Comment = await commentRepository.findOne({
@@ -21,22 +24,24 @@ export default async (request: Request, response: Response) => {
 			},
 		});
 
+		if (findComment.writer !== user.name || !user.is_admin) {
+			ColorConsole.red(`[ERROR 403] 댓글을 삭제할 권한이 없습니다.`);
+			handleFailed(response, 403, '댓글을 삭제할 권한이 없습니다.');
+			return;
+		}
+
 		if (!findComment) {
-			return response.status(404).json({
-				status: 404,
-				message: '존재하지 않는 댓글입니다.',
-			});
+			ColorConsole.red(`[ERROR 404] 존재하지 않는 댓글입니다.`);
+			handleFailed(response, 404, '존재하지 않는 댓글입니다.');
 		}
 
 		await commentRepository.remove(findComment);
-		return response.status(200).json({
-			status: 200,
-			message: '댓글 삭제를 성공하였습니다.',
-		});
+		ColorConsole.green(`[200] 댓글 삭제를 성공하였습니다.`);
+		handleSuccess(response, 200, '댓글 삭제를 성공하였습니다.');
+		return;
 	} catch (error) {
-		return response.status(500).json({
-			status: 500,
-			message: '서버 오류입니다.',
-		});
+		ColorConsole.red(`[ERROR 500] 서버 오류입니다. ${error.message}`);
+		handleFailed(response, 500, '서버 오류입니다.');
+		return;
 	}
 };
